AuthAPICorrect fix: create some widgets and add them to pages, fix styles
