+++ src/pages/404/index.tsx
@@ -1,32 +1,15 @@
 import { getRootLayout } from '@/shared/layouts'
-import { useTranslation } from '@/shared/lib/hooks'
-import { Button, HeadMeta, Typography } from '@/shared/ui'
-import Image from 'next/image'
-import Link from 'next/link'
+import { HeadMeta } from '@/shared/ui'
+import { NotFound } from '@/widgets/NotFound'
 
-import s from './NotFound.module.scss'
-
-const NotFound = () => {
-  const { text } = useTranslation()
-  const t = text.pages.notFound
-
+const NotFoundPage = () => {
   return (
     <>
       <HeadMeta title={'404 Not found'} />
-      <main className={s.wrapper}>
-        <Image alt={'not-found'} height={192} src={'/notFound.png'} width={451} />
-        <div>
-          <Typography className={s.title} variant={'regularText16'}>
-            {t.title}
-          </Typography>
-        </div>
-        <Button as={Link} fullWidth={false} href={'/'}>
-          {t.captionBtn}
-        </Button>
-      </main>
+      <NotFound />
     </>
   )
 }
 
-NotFound.getLayout = getRootLayout
-export default NotFound
+NotFoundPage.getLayout = getRootLayout
+export default NotFoundPage
