+++ src/pages/forgot-password/ForgotPasswordPage.module.scss
@@ -1,3 +1,3 @@
 .root {
-  margin-top: 36px;
+  margin: 36px;
 }
