+++ src/widgets/ConfirmedEmail/ui/ConfirmedEmail.module.scss
@@ -1,6 +1,7 @@
 .wrapper {
   width: 100%;
   max-width: 425px;
+  margin: 0 auto;
   text-align: center;
 
   .title {
