+++ src/widgets/EmailVerification/ui/index.ts
@@ -1,0 +1,1 @@
+export { EmailVerification } from 'src/widgets/EmailVerification/ui/EmailVerification'
