+++ src/widgets/ExpiredLink/ui/ExpiredLink.module.scss
@@ -1,6 +1,7 @@
 .wrapper {
   width: 100%;
   max-width: 475px;
+  margin: 0 auto;
   text-align: center;
 
   .title {
