+++ src/widgets/ForgotPassword/ui/ForgotPassword.module.scss
@@ -1,6 +1,12 @@
 .card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
   width: 100%;
   max-width: 378px;
+  margin: 0 auto;
   padding: 24px;
 }
