+++ src/widgets/NotFound/ui/NotFound.module.scss
