+++ src/widgets/NotFound/ui/index.ts
@@ -1,0 +1,1 @@
+export { NotFound } from './NotFound'
