+++ src/widgets/PrivacyPolicy/ui/PrivacyPolicy.module.scss
