+++ src/widgets/PrivacyPolicy/ui/index.ts
@@ -1,0 +1,1 @@
+export { PrivacyPolicy } from './PrivacyPolicy'
