+++ src/widgets/TermsOfService/index.ts
@@ -1,0 +1,1 @@
+export * from './ui'
