+++ src/widgets/TermsOfService/ui/TermsOfService.module.scss
