+++ src/widgets/TermsOfService/ui/index.ts
@@ -1,0 +1,1 @@
+export { TermsOfService } from './TermsOfService'
